src/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
langchain-openai==0.0.2
openai==1.6.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
# 许可证管理系统依赖
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pyjwt==2.8.0

# 多媒体处理依赖
boto3==1.34.0
botocore==1.34.0
celery==5.3.6
pillow==10.2.0
python-magic==0.4.27

# 文档处理依赖
pdfminer.six==20231228
python-docx==0.8.11
python-pptx==0.6.23
markdown==3.5.2
weasyprint==60.0

# 3D模型处理依赖
trimesh==4.0.8
numpy-stl==3.1.0
pygltflib==1.16.2

# AI代码补全系统依赖
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
sentence-transformers==2.2.2
faiss-cpu==1.9.0

# 联邦学习隐私保护系统依赖
# pysyft==0.8.1  # 暂时注释，该版本不支持 Python 3.12
torch==2.2.0
torchvision==0.17.0
tensorflow==2.16.2
opacus==1.5.4  # 差分隐私
kubernetes==28.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0

# Admin 后台设置管理依赖
asyncpg==0.29.0  # PostgreSQL 异步驱动
psycopg2-binary==2.9.9  # PostgreSQL 同步驱动
paho-mqtt==1.6.1  # MQTT 客户端
aiohttp==3.9.1  # 异步 HTTP 客户端（升级版本）
